APPLINK-4741: fixed some flexelint errors and warning

+++ src/components/application_manager/src/commands/hmi/on_button_press_notification.cc
@@ -55,7 +55,7 @@ void OnButtonPressNotification::Run() {
   event.set_smart_object(*message_);
   event.raise();
   (*message_)[strings::params][strings::function_id] =
-      mobile_apis::FunctionID::eType::OnButtonPressID;
+      static_cast<int>(mobile_apis::FunctionID::eType::OnButtonPressID);
   SendNotificationToMobile(message_);
 }
